Inventory System for Unity
Features

Can pick-up items off the ground if the item is set to auto-pick-up.
Otherwise, using an interaction system and implementing the IInteractable interface, allows the player to collide with and prompt the player for item pick-up.

Hotbar inventory with mouse-scroll wheel controls and highlighting.

Options for player backpack for extended storage.

Options for chest inventory for objects in world which hold an inventory. Can also be used for felled enemies.

Item Data for saving things like transform positions, whether or not the item has already been collected, and things of that nature.

Items can be picked up via the mouse and moved to an open slot or swapped with an existing item.

Inventory holder is placed on the character.

Inventory slots can be adjusted for hotbar, backpack, etc.